Migrated tabata to the new end structure.
